rmics/requirements.txt
asgiref==3.6.0
Django==4.1.7
mysqlclient==2.1.1
Pillow==9.5.0
sqlparse==0.4.3
tzdata==2022.7
python-decouple==3.8
django-notifications-hq==1.6.0
